Add function for finding old inboxes
This function only checks if an inbox is older than the current time.
What this means is that some inboxes might only live for a short time if
they are created right before the function is run and not a full 24
hours.

An improvement in the future might be to update this function to check
for inboxes that are older than 24 hours, and run that function evvery
hour or so.